timers: Fix Sire stun timer duration
Commit 604d9d9fd20d8c7e5f481b0222826fb103631091 originally attempts to
fix this timer, but was using the wrong starting tick in its
calculations. (That is, it used the idle -> wakeup change tick instead
of the awakened -> stunned change tick) This led to the timer still
being too long, because of the 2 additional ticks at the start of the
fight.

(cherry picked from commit b9342a01e5326f6acb4ebb7c10da748bff23e4a8)