screenshot: draw report button date at correct location
This commit reworks the report button date option to change the widget
text rather than drawing an overlay, which ensures the text is still
drawn correctly even if the chatbox is moved. As a bonus, because it is
not drawn as an overlay, the overlay class (and all related
dependencies) can be removed.

(cherry picked from commit 0be27f3eec4481498639e4024b8bc4a25cfb600c)